add: fullsync and unit test
