+++ include/DataObject.h
@@ -91,11 +91,11 @@ public:
     friend ostream& operator<<(ostream& out, const DataObject &datum);
 
     // comparisons
-    bool operator < (const DataObject second) const{
+    bool operator < (const DataObject& second) const{
         return this->myBuffer < second.to_ZZ();
     }
 
-    bool operator== (const DataObject second) const {
+    bool operator== (const DataObject& second) const {
         return this->myBuffer == second.to_ZZ();
     }
