+++ include/InterCPISync.h
@@ -79,8 +79,6 @@ public:
      */
     bool SyncServer(shared_ptr<Communicant> commSync, list<DataObject*> &selfMinusOther, list<DataObject*> &otherMinusSelf);
 
-    InterCPISync* clone() const;
-
     /*
      ** update metadata when an element is being added.
      * The maximum number of elements that the data structure can store is 1<<bitNum
