+++ include/Logger.h
@@ -40,7 +40,7 @@ public:
     };
     
     // Variables
-    static const LOG_TYPES logLevel = COMM_DETAILS; /** A number representing the level logging desired.  All logs of level <= logLevel are outputted.*/
+    static const LOG_TYPES logLevel = NONE; /** A number representing the level logging desired.  All logs of level <= logLevel are outputted.*/
     
     /**
      * Default constructor
