+++ tests/sys/long/GenSyncTest.cpp
@@ -14,6 +14,7 @@
 #include "GenSync.h"
 #include "CommDummy.h"
 #include "Auxiliary.h"
+#include "TestAuxiliary.h"
 
 
 // constants
@@ -29,289 +30,6 @@ void GenSyncTest::setUp(){
     srand(SEED);
 }
 void GenSyncTest::tearDown(){}
-
-// helpers
-
-vector<GenSync> GenSyncTest::builderCombos() {
-    vector<GenSync> ret;
-
-    // iterate through all possible combinations of communicant and syncmethod
-    for(auto prot = GenSync::SyncProtocol::BEGIN; prot != GenSync::SyncProtocol::END; ++prot) {
-        for(auto comm = GenSync::SyncComm::BEGIN; comm != GenSync::SyncComm::END; ++comm) {
-            GenSync::Builder builder = GenSync::Builder().
-                    setProtocol(prot).
-                    setComm(comm);
-
-            switch(prot) {
-                case GenSync::SyncProtocol::CPISync:
-                case GenSync::SyncProtocol::OneWayCPISync:
-                    builder.
-                            setBits(eltSize).
-                            setMbar(mBar);
-                    break;
-                case GenSync::SyncProtocol::InteractiveCPISync:
-                    builder.
-                            setBits(eltSize).
-                            setMbar(mBar).
-                            setNumPartitions(numParts);
-                    break;
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            switch(comm) {
-                case GenSync::SyncComm::string:
-                    builder.
-                            setIoStr(iostr);
-                    break;
-                case GenSync::SyncComm::socket:
-                    builder.
-                            setHost(host).
-                            setPort(port);
-                    break;
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            ret.emplace_back(GenSync(builder.build()));
-        }
-    }
-    return ret;
-}
-
-vector<GenSync> GenSyncTest::constructorCombos(bool useFile) {
-    vector<GenSync> ret;
-
-    // iterate through all possible combinations of communicant and syncmethod
-    for(auto prot = GenSync::SyncProtocol::BEGIN; prot != GenSync::SyncProtocol::END; ++prot) {
-        for(auto comm = GenSync::SyncComm::BEGIN; comm != GenSync::SyncComm::END; ++comm) {
-            vector<shared_ptr<Communicant>> communicants;
-            vector<shared_ptr<SyncMethod>> methods;
-            switch(prot) {
-                case GenSync::SyncProtocol::CPISync:
-                    methods = {make_shared<ProbCPISync>(mBar, eltSize, err)};
-                    break;
-                case GenSync::SyncProtocol::OneWayCPISync:
-                    methods = {make_shared<CPISync_HalfRound>(mBar, eltSize, err)};
-                    break;
-                case GenSync::SyncProtocol::InteractiveCPISync:
-                    methods = {make_shared<InterCPISync>(mBar, eltSize, err, numParts)};
-                    break;
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            switch(comm) {
-                case GenSync::SyncComm::string:
-                    communicants = {make_shared<CommString>(iostr, b64)};
-                    break;
-                case GenSync::SyncComm::socket:
-                    communicants = {make_shared<CommSocket>(port, host)};
-                    break;
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            // call constructor depending on useFile
-            ret.emplace_back(useFile ? GenSync(communicants, methods, temporaryDir() + "/gensynctest/" + toStr(rand())) : GenSync(communicants, methods));
-        }
-    }
-    return ret;
-}
-
-vector<GenSync> GenSyncTest::oneWayCombos() {
-    vector<GenSync> ret;
-
-    // iterate through all possible combinations of communicant and syncmethod
-    for(auto prot = GenSync::SyncProtocol::BEGIN; prot != GenSync::SyncProtocol::END; ++prot) {
-        for(auto comm = GenSync::SyncComm::BEGIN; comm != GenSync::SyncComm::END; ++comm) {
-            vector<shared_ptr<Communicant>> communicants;
-            vector<shared_ptr<SyncMethod>> methods;
-            switch(prot) {
-                case GenSync::SyncProtocol::CPISync:
-                    methods = {make_shared<ProbCPISync>(mBar, eltSize, err)};
-                    break;
-                case GenSync::SyncProtocol::OneWayCPISync:
-                    methods = {make_shared<CPISync_HalfRound>(mBar, eltSize, err)};
-                case GenSync::SyncProtocol::InteractiveCPISync:
-                    methods = {make_shared<InterCPISync>(mBar, eltSize, err, numParts)};
-                    break;
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            switch(comm) {
-                case GenSync::SyncComm::socket:
-                    communicants = {make_shared<CommSocket>(port, host)};
-                    break;
-                case GenSync::SyncComm::string:
-                    continue; // not a 1way sync
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-            ret.push_back(GenSync(communicants, methods));
-        }
-    }
-    return ret;
-}
-
-vector<GenSync> GenSyncTest::twoWayCombos() {
-    vector<GenSync> ret;
-
-    // iterate through all possible combinations of communicant and syncmethod
-    for(auto prot = GenSync::SyncProtocol::BEGIN; prot != GenSync::SyncProtocol::END; ++prot) {
-        for(auto comm = GenSync::SyncComm::BEGIN; comm != GenSync::SyncComm::END; ++comm) {
-            vector<shared_ptr<Communicant>> communicants;
-            vector<shared_ptr<SyncMethod>> methods;
-            switch(prot) {
-                case GenSync::SyncProtocol::CPISync:
-                    methods = {make_shared<ProbCPISync>(mBar, eltSize, err)};
-                    break;
-                case GenSync::SyncProtocol::OneWayCPISync:
-                    continue; // not a 2way sync
-                case GenSync::SyncProtocol::InteractiveCPISync:
-                    methods = {make_shared<InterCPISync>(mBar, eltSize, err, numParts)};
-                    break;
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            switch(comm) {
-                case GenSync::SyncComm::socket:
-                    communicants = {make_shared<CommSocket>(port, host)};
-                    break;
-                case GenSync::SyncComm::string:
-                    continue; // not a 2way sync
-                default:
-                    CPPUNIT_FAIL("Unreachable");
-                    break;
-            }
-
-            ret.push_back(GenSync(communicants, methods));
-        }
-    }
-    return ret;
-}
-
-vector<GenSync> GenSyncTest::standardCombos() {
-    return constructorCombos(false);
-}
-
-vector<GenSync> GenSyncTest::fileCombos() {
-    return constructorCombos(true);
-}
-
-void GenSyncTest::_syncTest(bool oneWay, GenSync GenSyncServer, GenSync GenSyncClient) {
-    for(int jj = 0; jj < NUM_TESTS; jj++) {
-        // setup DataObjects
-        const unsigned char SIMILAR = randByte(); // amt of elems common to both GenSyncs
-        const unsigned char CLIENT_MINUS_SERVER = randByte(); // amt of elems unique to client
-        const unsigned char SERVER_MINUS_CLIENT = randByte(); // amt of elems unique to server
-
-        vector<DataObject *> objectsPtr;
-
-        for (unsigned long ii = 0; ii < SIMILAR + SERVER_MINUS_CLIENT + CLIENT_MINUS_SERVER -
-                                        1; ii++) {
-            objectsPtr.push_back(new DataObject(randZZ()));
-        }
-        ZZ *last = new ZZ(randZZ()); // last datum represented by a ZZ so that the templated addElem can be tested
-        objectsPtr.push_back(new DataObject(*last));
-
-        // ... add data objects unique to the server
-        for (auto iter = objectsPtr.begin(); iter != objectsPtr.begin() + SERVER_MINUS_CLIENT; iter++) {
-            GenSyncServer.addElem(*iter);
-        }
-
-        // ... add data objects unique to the client
-        for (auto iter = objectsPtr.begin() + SERVER_MINUS_CLIENT;
-             iter != objectsPtr.begin() + SERVER_MINUS_CLIENT + CLIENT_MINUS_SERVER; iter++) {
-            GenSyncClient.addElem(*iter);
-        }
-
-        // add common data objects to both
-        for (auto iter = objectsPtr.begin() + SERVER_MINUS_CLIENT + CLIENT_MINUS_SERVER;
-             iter != objectsPtr.end() - 1; iter++) { // minus 1 so that the templated element can be tested
-            GenSyncClient.addElem(*iter);
-            GenSyncServer.addElem(*iter);
-        }
-
-        // ensure that adding a object that fits the generic type T works
-        GenSyncClient.addElem(last);
-        GenSyncServer.addElem(last);
-
-        // create the expected reconciled multiset
-        multiset<string> reconciled;
-        for (auto dop : objectsPtr) {
-            reconciled.insert(dop->print());
-        }
-
-        // create two processes to test successful reconciliation. the parent process tests the server; the child, the client.
-        int err = 1;
-        int chld_state;
-        int my_opt = 0;
-        pid_t pID = fork();
-        if (pID == 0) {
-            signal(SIGCHLD, SIG_IGN);
-            if(!oneWay) {
-                // reconcile client with server
-                forkHandleReport resultClient = forkHandle(GenSyncClient, GenSyncServer);
-
-                // check reasonable statistics
-                CPPUNIT_ASSERT(resultClient.bytes > 0);
-                CPPUNIT_ASSERT(resultClient.success);
-
-                // convert reconciled elements into string representation
-                multiset<string> resClient;
-                for (auto dop : GenSyncClient.dumpElements()) {
-                    resClient.insert(dop->print());
-                }
-
-                // check that expected and resultant reconciled sets match up in both size and contents
-                CPPUNIT_ASSERT_EQUAL(reconciled.size(), resClient.size());
-                CPPUNIT_ASSERT(multisetDiff(reconciled, resClient).empty());
-            }
-            exit(0);
-        } else if (pID < 0) {
-            cout << "throw out err = " << err << endl;
-            throw err;
-        } else {
-            // wait for child process to complete
-            waitpid(pID, &chld_state, my_opt);
-
-            // reconcile server with client
-            forkHandleReport resultServer = forkHandle(GenSyncServer, GenSyncClient);
-
-            // check reasonable statistics
-            CPPUNIT_ASSERT(resultServer.bytes > 0);
-            CPPUNIT_ASSERT(resultServer.success);
-
-            // convert reconciled elements into string representation
-            multiset<string> resServer;
-            for (auto dop : GenSyncServer.dumpElements()) {
-                resServer.insert(dop->print());
-            }
-
-            // check that expected and resultant reconciled sets match up in both size and contents
-            CPPUNIT_ASSERT_EQUAL(reconciled.size(), resServer.size());
-            CPPUNIT_ASSERT(multisetDiff(reconciled, resServer).empty());
-        }
-    }
-}
-
-void GenSyncTest::syncTestOneWay(GenSync GenSyncClient, GenSync GenSyncServer) {
-    _syncTest(true, GenSyncClient, GenSyncServer);
-}
-
-void GenSyncTest::syncTest(GenSync GenSyncServer, GenSync GenSyncClient) {
-    _syncTest(false, GenSyncServer, GenSyncClient);
-}
 
 // tests
