+++ tests/sys/long/GenSyncTest.h
@@ -52,79 +52,6 @@ public:
 
     // Tests getPort
     void testPort();
-private:
-    // constants
-    const int NUM_TESTS = 1; // Times to run oneWay and twoWay sync tests
-
-    const size_t eltSize = sizeof(randZZ()) * CHAR_BIT; // size of elements stored in sync tests
-    const int mBar = UCHAR_MAX*2; // max differences between client and server in sync tests
-    const string iostr = ""; // initial string used to construct CommString
-    const bool b64 = true; // whether CommString should communicate in b64
-    const string host = "localhost"; // host for CommSocket
-    const unsigned int port = 8031; // port for CommSocket
-    const int err = 8; // negative log of acceptable error probability for probabilistic syncs
-    const int numParts = 3; // partitions per level for divide-and-conquer syncs
-
-    // helpers
-
-    /**
-     * returns all possible gensync configurations constructed using GenSync::Builder
-     * GenSyncs are lexicographically ordered with index of SyncProtocol more significant than the index of SyncComm
-     */
-    vector<GenSync> builderCombos();
-
-    /**
-     * return a vector of GenSyncs constructed using either of the two GenSync constructors
-     * GenSyncs are lexicographically ordered with index of SyncProtocol more significant than the index of SyncComm
-     * @param useFile if true, GenSyncs constructed using file-based constructor. otherwise, constructed using other constructor
-     */
-    vector<GenSync> constructorCombos(bool useFile);
-
-    /**
-     * returns gensync configurations that communicate one-way constructed using the standard way of creating GenSync objects (without a file)
-     * GenSyncs are lexicographically ordered with index of SyncProtocol more significant than the index of SyncComm
-     */
-    vector<GenSync> oneWayCombos();
-
-    /**
-     * returns gensync configurations that communicate two ways constructed using the standard way of creating GenSync objects (without a file)
-     * GenSyncs are lexicographically ordered with index of SyncProtocol more significant than the index of SyncComm
-     */
-    vector<GenSync> twoWayCombos();
-
-    /**
-     * returns all possible gensync configurations constructed using the standard way of creating GenSync objects (without a file)
-     * GenSyncs are lexicographically ordered with index of SyncProtocol more significant than the index of SyncComm
-     */
-    vector<GenSync> standardCombos();
-
-    /**
-     * returns all possible gensync configurations constructed by creating GenSync objects using a file to store contents
-     * GenSyncs are lexicographically ordered with index of SyncProtocol more significant than the index of SyncComm
-     */
-    vector<GenSync> fileCombos();
-
-    /**
-     * Runs tests assuring that two GenSync objects successfully sync via two-way communication
-     * @param oneWay true iff the sync will be one way (only server is reconciled)
-     * @param GenSyncServer Server GenSync
-     * @param GenSyncClient Client GenSync
-     */
-    void _syncTest(bool oneWay, GenSync GenSyncServer, GenSync GenSyncClient);
-
-    /**
-     * One way synctest
-     * @param GenSyncClient
-     * @param GenSyncServer
-     */
-    void syncTestOneWay(GenSync GenSyncClient, GenSync GenSyncServer);
-
-    /**
-     * Two way synctest
-     * @param GenSyncServer Server GenSync
-     * @param GenSyncClient Client GenSync
-     */
-    void syncTest(GenSync GenSyncClient, GenSync GenSyncServer);
 
 };
